Fix all warnings: zero -Wall warnings across 42 source files
- sdlog.h: __attribute__((unused)) on all 13 static functions
- Format strings: %llu->%lu, %lld->%ld for AArch64 LP64 model (11 files)
- inference.c: 500000ULL->500000UL literal type fix
- Unused variables: (void)cast suppression across 12 files
- Unused functions: __attribute__((unused)) on 14 static functions
- aishell.c: fix reversed git_repo_init() arguments (pointer type bug)
- build_rpi.ps1: replace -Wno-* suppressions with -Wall (all clean)

+++ kernel/core/main.c
@@ -872,10 +872,10 @@ static void kernel_idle_loop(void)
             kstate.uptime_ticks++;
 
             /* Log health status */
-            kprintf_debug("[IDLE] ops=%llu models=%u mem_free=%u MB\n",
+            kprintf_debug("[IDLE] ops=%lu models=%u mem_free=%lu MB\n",
                          kstate.tensor_ops_total,
                          kstate.models_loaded,
-                         tensor_mm_free_bytes() / (1024 * 1024));
+                         (uint64_t)(tensor_mm_free_bytes() / (1024 * 1024)));
         }
 
         idle_cycles++;
